feat: 존재하는 예약에 해당하는 시간 삭제할 때 예외 처리 추가

+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -2,9 +2,11 @@ package roomescape.reservationtime.service;
 
 import jakarta.validation.Valid;
 import java.util.List;
+import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.reservationtime.dao.ReservationTimeDao;
 import roomescape.reservationtime.dto.request.ReservationTimeRequest;
@@ -14,9 +16,11 @@ import roomescape.reservationtime.dto.response.ReservationTimeResponse;
 public class ReservationTimeService {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<ReservationTimeResponse> findAll() {
@@ -35,6 +39,12 @@ public class ReservationTimeService {
     }
 
     public Integer delete(@PathVariable Long id) {
+        if (reservationTimeDao.findById(id).isEmpty()) {
+            throw new NoSuchElementException("예약 시간이 존재하지 않습니다.");
+        }
+        if (reservationDao.findByTimeId(id).isPresent()) {
+            throw new IllegalStateException("해당 시간에 대한 예약이 존재합니다.");
+        }
         return reservationTimeDao.delete(id);
     }
 }
